update env example with keys
