+++ .env.example
@@ -1,4 +1,4 @@
 CLOUDINARY_CLOUD_NAME=
-CLOUDINARY_API_KEY=
+CLOUDINARY_API_KEY=983753278431369
 CLOUDINARY_API_SECRET=
-NEXT_PUBLIC_LIVEBLOCKS_PUBLIC_KEY=
+NEXT_PUBLIC_LIVEBLOCKS_PUBLIC_KEY=pk_dev_TyJoLcNht-pFDqJxctnpShxkV4xgYWWVcZMibBSthGiESihddBYvPocQp3qZBFQq
